% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getMap}
\alias{getMap}
\title{Extract the testmap Field of a Foo Class Instance}
\usage{
getMap(instance)
}
\arguments{
\item{instance}{An instance of the Foo class.}
}
\description{
This function takes an instance of the Foo class and returns
the testmap field. This is possible due to: 1) exposing Foo by setting
RCPP_EXPOSED_CLASS(Foo) in RcppExtModEx_types.h, and 2) using the
custom wrap handler defined in RcppExtModEx_types.cpp and RcppExtModEx_types.h.
}
