% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{makeFoo}
\alias{makeFoo}
\title{Create an Instance of the Foo Class}
\usage{
makeFoo()
}
\description{
This function demonstrates how a Foo object is created and
returned to R.
}
\details{
Using Foo as C++ return value is possible due to exposing the class
with RCPP_EXPOSED_CLASS(Foo) in RcppExtModEx_types.h.
}
