% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mapIn_and_mapOut}
\alias{mapIn_and_mapOut}
\title{Provide a Matrix and Return a Matrix}
\usage{
mapIn_and_mapOut(mymap)
}
\arguments{
\item{mymap}{A Matrix with column names. The column names are used as keys
in the C++ map.}
}
\description{
This function demonstrates how a Matrix is passed to C++ as
a map. And how the C++ map is returned to R
as Matrix.
}
\details{
Conversion from R Matrix to C++ std::map<std::string, arma::colvec>
is handled by Rcpp as and wrap functions defined in RcppExtModEx_types.cpp
and RcppExtModEx_types.h.
}
