% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setMap}
\alias{setMap}
\title{Set the testmap Field of a Foo Class Instance}
\usage{
setMap(instance, mymap)
}
\arguments{
\item{instance}{An instance of the Foo class.}

\item{mymap}{A Matrix with column names. The column names are used as keys
in the C++ map.}
}
\description{
This function takes an instance of the Foo class assigns mymap
to the testmap field.
}
\details{
This is possible due to: 1) exposing Foo by
setting RCPP_EXPOSED_CLASS(Foo) in RcppExtModEx_types.h, and 2) using
the custom as handler defined in RcppExtModEx_types.cpp and
RcppExtModEx_types.h.
}
